requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
Django==4.0.3
django-phonenumber-field==6.1.0
django-silk==4.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
gprof2dot==2021.2.21
idna==3.3
Jinja2==3.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.45
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
